Change layout to reference the previous sibling
This changes how the layout tree is linked and the protocol for
computing layout. Basically, in addition to a parent and a list of
children, each layout object now also has a pointer to its previous
sibling. That way, to compute its Y-position, a layout object can
reference its previous sibling's Y-position and height.

This might seem more complex than having the parent keep track of a
vertical cursor for children, but the advantage of this approach is
that computing the layout of each layout object is totally
self-contained to a single `layout` function, instead of
responsibilities being split between that node's and its parent's
`layout` functions. It also means that the recursive calls to
`layout` are purely that---calling `layout` on each child---instead of
being interlaced with code that tracks a vertical cursor.