Replace `has_block_children` with something that defaults to `block`
This isn't a full implementation of anonymous block boxes because in
cases like this

    <p>This <span>is</span> <div>a block</div> inside a paragraph</p>

it will do the wrong thing---"This" and "is" will be split into two
paragraphs instead of being placed in one. But it's pretty close and,
importantly, won't default to inline layout mode in a way that
confused students last year.