Fix chapter 13
This one was a little tricky. This lab copies in all the layout types
for two reasons. First, it used to add support for width/height; it
doesn't any more, so I actually stripped that out in this change.
Second, it redefines DrawRect and friends; for that reason I copied in
the lab11 copies of the layout types.