Unify InlineLayout and BlockLayout (#659)
* Unify InlineLayout and BlockLayout

* Update code for chapters 5-10 to unify Block/InlineLayout

* Update chapter 11

* Fix chapter 12 (easy)

* Fix chapter 13

This one was a little tricky. This lab copies in all the layout types
for two reasons. First, it used to add support for width/height; it
doesn't any more, so I actually stripped that out in this change.
Second, it redefines DrawRect and friends; for that reason I copied in
the lab11 copies of the layout types.

* Fix up Chapter 13

* Fix Chapter 14

This required a change to the tests because I also removed support for
width/height (which was in the code but actually gone from the chapter
text).

* Fix text of every chapter except 5 and 7

* Fix up chapter 7 (and more fixes to Chapter 8)

* Make cursor_x and cursor_y relative throughout

* Write a new version of Chapter 5 with the unified BlockLayout

* Add any() and all() support to compiler

* More fixes for uniformity and tests passing

* Update GH Action so it works

* Tweaks from @chrishtr

* Fix redundancies between two paragraphs

* Tweaks

* Fix widget

* Add feature to hide certain methods from JS compiler

* Fix widget size

* Typo

* Avoid scroll bars in widgets when input controls are expanded