Split the "interactive widgets" section up
The discussion of plugins went into a Go Further block, which itself
merged with a different Go Further block. The discussion of canvas
moved into a footnote, as did the discussion of input element
customization. (They have an exercise & a go further in a prior
chapter.) Compressing the remainder left a paragraph that got moved
into the next section.