Split the "interactive widgets" section up (#723)
* Split the "interactive widgets" section up

The discussion of plugins went into a Go Further block, which itself
merged with a different Go Further block. The discussion of canvas
moved into a footnote, as did the discussion of input element
customization. (They have an exercise & a go further in a prior
chapter.) Compressing the remainder left a paragraph that got moved
into the next section.

* Update book/embeds.md

* Clarify PDF standard, and improve phrasing about customized form controls

---------

Co-authored-by: Chris Harrelson <3453258+chrishtr@users.noreply.github.com>
Co-authored-by: Chris Harrelson <chrishtr@gmail.com>